refactor(core)!: email ops on KMPAuthBackend, FirebaseEmailAuth removed
Two API styles coexisted: sign-in states plus a Firebase-specific
FirebaseEmailAuth object for the non-launchable email operations. Those
operations are not Firebase-specific concepts, so they now live on the
provider-agnostic AuthProviderBackend (default: unsupported failure, so
custom backends stay source-compatible):

- sendPasswordResetEmail(email, actionCodeSettings)
- sendSignInLinkToEmail(email, actionCodeSettings)
- isSignInWithEmailLink(link) / signInWithEmailLink(email, link, linkAccount)

EmailActionCodeSettings moved to kmpauth-core. FirebaseAuthBackend
implements the operations via GitLive; the wasm actual reports the wasm
reason.

KMPAuthBackend now implements AuthProviderBackend itself, delegating to
the registered backend - callers use KMPAuthBackend.sendPasswordResetEmail(...)
or KMPAuthBackend.reauthenticate(...) directly instead of the
KMPAuthBackend.require() indirection; with no backend registered,
Result-returning operations fail with the how-to-register message instead
of throwing. require()/getOrNull() remain for code that wants the raw
backend.

FirebaseEmailAuth (unreleased, added earlier in this PR) is deleted.
