+++ backends/firebase/kmpauth-firebase-core/api/jvm/kmpauth-firebase-core.api
@@ -15,8 +15,12 @@ public final class com/mmk/kmpauth/firebase/backend/FirebaseAuthBackend : com/mm
 	public static final field $stable I
 	public static final field INSTANCE Lcom/mmk/kmpauth/firebase/backend/FirebaseAuthBackend;
 	public fun currentUser ()Lcom/mmk/kmpauth/core/auth/KMPAuthUser;
+	public fun isSignInWithEmailLink (Ljava/lang/String;)Z
 	public fun reauthenticate-gIAlu-s (Lcom/mmk/kmpauth/core/auth/AuthCredential;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+	public fun sendPasswordResetEmail-0E7RQCE (Ljava/lang/String;Lcom/mmk/kmpauth/core/auth/EmailActionCodeSettings;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+	public fun sendSignInLinkToEmail-0E7RQCE (Ljava/lang/String;Lcom/mmk/kmpauth/core/auth/EmailActionCodeSettings;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 	public fun signIn-0E7RQCE (Lcom/mmk/kmpauth/core/auth/AuthCredential;ZLkotlin/coroutines/Continuation;)Ljava/lang/Object;
+	public fun signInWithEmailLink-BWLJW6A (Ljava/lang/String;Ljava/lang/String;ZLkotlin/coroutines/Continuation;)Ljava/lang/Object;
 	public fun signOut (Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 }
 
@@ -31,49 +35,12 @@ public final class com/mmk/kmpauth/firebase/backend/FirebaseKMPAuthUser : com/mm
 	public fun getUid ()Ljava/lang/String;
 }
 
-public final class com/mmk/kmpauth/firebase/email/EmailActionCodeSettings {
-	public static final field $stable I
-	public fun <init> (Ljava/lang/String;ZLjava/lang/String;Ljava/lang/String;ZLjava/lang/String;Ljava/lang/String;)V
-	public synthetic fun <init> (Ljava/lang/String;ZLjava/lang/String;Ljava/lang/String;ZLjava/lang/String;Ljava/lang/String;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public final fun component1 ()Ljava/lang/String;
-	public final fun component2 ()Z
-	public final fun component3 ()Ljava/lang/String;
-	public final fun component4 ()Ljava/lang/String;
-	public final fun component5 ()Z
-	public final fun component6 ()Ljava/lang/String;
-	public final fun component7 ()Ljava/lang/String;
-	public final fun copy (Ljava/lang/String;ZLjava/lang/String;Ljava/lang/String;ZLjava/lang/String;Ljava/lang/String;)Lcom/mmk/kmpauth/firebase/email/EmailActionCodeSettings;
-	public static synthetic fun copy$default (Lcom/mmk/kmpauth/firebase/email/EmailActionCodeSettings;Ljava/lang/String;ZLjava/lang/String;Ljava/lang/String;ZLjava/lang/String;Ljava/lang/String;ILjava/lang/Object;)Lcom/mmk/kmpauth/firebase/email/EmailActionCodeSettings;
-	public fun equals (Ljava/lang/Object;)Z
-	public final fun getAndroidInstallIfNotAvailable ()Z
-	public final fun getAndroidMinimumVersion ()Ljava/lang/String;
-	public final fun getAndroidPackageName ()Ljava/lang/String;
-	public final fun getCanHandleCodeInApp ()Z
-	public final fun getIOSBundleId ()Ljava/lang/String;
-	public final fun getLinkDomain ()Ljava/lang/String;
-	public final fun getUrl ()Ljava/lang/String;
-	public fun hashCode ()I
-	public fun toString ()Ljava/lang/String;
-}
-
 public final class com/mmk/kmpauth/firebase/email/EmailAuthMode : java/lang/Enum {
 	public static final field SignIn Lcom/mmk/kmpauth/firebase/email/EmailAuthMode;
 	public static final field SignUp Lcom/mmk/kmpauth/firebase/email/EmailAuthMode;
 	public static fun getEntries ()Lkotlin/enums/EnumEntries;
 	public static fun valueOf (Ljava/lang/String;)Lcom/mmk/kmpauth/firebase/email/EmailAuthMode;
 	public static fun values ()[Lcom/mmk/kmpauth/firebase/email/EmailAuthMode;
-}
-
-public final class com/mmk/kmpauth/firebase/email/FirebaseEmailAuth {
-	public static final field $stable I
-	public static final field INSTANCE Lcom/mmk/kmpauth/firebase/email/FirebaseEmailAuth;
-	public final fun isSignInWithEmailLink (Ljava/lang/String;)Z
-	public final fun reauthenticate-0E7RQCE (Ljava/lang/String;Ljava/lang/String;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public final fun sendPasswordResetEmail-0E7RQCE (Ljava/lang/String;Lcom/mmk/kmpauth/firebase/email/EmailActionCodeSettings;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public static synthetic fun sendPasswordResetEmail-0E7RQCE$default (Lcom/mmk/kmpauth/firebase/email/FirebaseEmailAuth;Ljava/lang/String;Lcom/mmk/kmpauth/firebase/email/EmailActionCodeSettings;Lkotlin/coroutines/Continuation;ILjava/lang/Object;)Ljava/lang/Object;
-	public final fun sendSignInLinkToEmail-0E7RQCE (Ljava/lang/String;Lcom/mmk/kmpauth/firebase/email/EmailActionCodeSettings;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public final fun signInWithEmailLink-BWLJW6A (Ljava/lang/String;Ljava/lang/String;ZLkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public static synthetic fun signInWithEmailLink-BWLJW6A$default (Lcom/mmk/kmpauth/firebase/email/FirebaseEmailAuth;Ljava/lang/String;Ljava/lang/String;ZLkotlin/coroutines/Continuation;ILjava/lang/Object;)Ljava/lang/Object;
 }
 
 public final class com/mmk/kmpauth/firebase/email/FirebaseEmailSignInStateKt {
