+++ backends/firebase/kmpauth-firebase-core/src/wasmJsMain/kotlin/com/mmk/kmpauth/firebase/backend/FirebaseAuthBackend.wasm.kt
@@ -2,6 +2,7 @@ package com.mmk.kmpauth.firebase.backend
 
 import com.mmk.kmpauth.core.auth.AuthCredential
 import com.mmk.kmpauth.core.auth.AuthProviderBackend
+import com.mmk.kmpauth.core.auth.EmailActionCodeSettings
 import com.mmk.kmpauth.core.auth.KMPAuthUser
 
 internal const val WASM_UNSUPPORTED_REASON: String =
@@ -17,6 +18,25 @@ public actual object FirebaseAuthBackend : AuthProviderBackend {
         Result.failure(UnsupportedOperationException(WASM_UNSUPPORTED_REASON))
 
     override suspend fun reauthenticate(credential: AuthCredential): Result<Unit> =
+        Result.failure(UnsupportedOperationException(WASM_UNSUPPORTED_REASON))
+
+    override suspend fun sendPasswordResetEmail(
+        email: String,
+        actionCodeSettings: EmailActionCodeSettings?,
+    ): Result<Unit> = Result.failure(UnsupportedOperationException(WASM_UNSUPPORTED_REASON))
+
+    override suspend fun sendSignInLinkToEmail(
+        email: String,
+        actionCodeSettings: EmailActionCodeSettings,
+    ): Result<Unit> = Result.failure(UnsupportedOperationException(WASM_UNSUPPORTED_REASON))
+
+    override fun isSignInWithEmailLink(link: String): Boolean = false
+
+    override suspend fun signInWithEmailLink(
+        email: String,
+        link: String,
+        linkAccount: Boolean,
+    ): Result<KMPAuthUser?> =
         Result.failure(UnsupportedOperationException(WASM_UNSUPPORTED_REASON))
 
     override suspend fun signOut(): Unit = Unit
